+++ src/Modules/TitelNormen.php
@@ -52,95 +52,141 @@ class TitelNormen extends \Module
 		$objMembers = \Database::getInstance()->prepare('SELECT * FROM tl_mitgliederverwaltung WHERE published = ?')
 		                                      ->execute(1);
 
-		// Aktuelle Titel und Normen suchen
 		$daten = array();
+		$titel = array();
+		$normen = array();
 		$mindate = date('Ymd', strtotime($this->mitgliederverwaltung_zeitraum)); // Nur Normen/Titel eines bestimmten Zeitraums
 		$maxdate = date('Ymd'); // Heutiges Datum setzen
 		
+		// Titel und Normen auslesen
 		if($objMembers->numRows)
 		{
 			while($objMembers->next())
 			{
-				$normen = unserialize($objMembers->normen); // Normen extrahieren
-				// Normen-Markierung erstellen, wird benötigt um für Titel benutzte Normen zu kennzeichnen
-				if(is_array($normen))
-				{
-					for($x = 0; $x < count($normen); $x++)
-					{
-						$normen[$x]['markiert'] = false;
-					}
-				}
 
 				// Titel der Reihe nach abfragen
-				foreach($GLOBALS['TL_LANG']['tl_mitgliederverwaltung']['normen_titel'] as $titel => $titelname)
+				foreach($GLOBALS['TL_LANG']['tl_mitgliederverwaltung']['normen_titel'] as $titelkey => $titelname)
 				{
-					// Normen dieses Titels suchen und merken
-					$aktuelleNorm = array(); // Nimmt die Daten der aktuellen Norm auf (Turniername, Link)
-					$aktuellerAbstand = 99999; // Zeitlichen Abstand vom Titeldatum zum Normdatum resetten
-
-					if($objMembers->{$titel.'_title'})
+					if($objMembers->{$titelkey.'_title'})
 					{
-						// Titel aktiv, jetzt Normen prüfen, die dazu gehören
-						if(is_array($normen))
-						{
-							for($x = 0; $x < count($normen); $x++)
-							{
-								if($normen[$x]['date'] >= $mindate && $normen[$x]['date'] <= $maxdate && $normen[$x]['title'] == $titel)
-								{
-									// Normdatum innerhalb des Zeitfensters und Normtitel entspricht erreichtem Titel
-									// Zeitlichen Abstand berechnen und Daten neu setzen, wenn Abstand kleiner ist
-									$abstand = $objMembers->{$titel.'_date'} - $normen[$x]['date'];
-									if($abstand < $aktuellerAbstand)
-									{
-										$aktuellerAbstand = $abstand;
-										$aktuelleNorm = array
-										(
-											'turnier'     => $normen[$x]['tournament'],
-											'link'        => $normen[$x]['url'],
-										);
-									}
-									$normen[$x]['markiert'] = true;
-                				}
-                			}
-                			// Alle Titelnormen geprüft, jetzt Daten zuweisen
-							$daten[$titel]['titel'][] = array
-							(
-								'nachname'    => $objMembers->nachname,
-								'vorname'     => $objMembers->vorname,
-								'name'        => $objMembers->vorname.' '.$objMembers->nachname,
-								'norm'        => $titel,
-								'datum'       => \Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($objMembers->{$titel.'_date'}),
-								'turnier'     => $aktuelleNorm['turnier'],
-								'link'        => $aktuelleNorm['link'],
-								'turnierlink' => $aktuelleNorm['link'] ? '<a href="'.$aktuelleNorm['link'].'" target="_blank">'.$aktuelleNorm['turnier'].'</a>' : $aktuelleNorm['turnier'],
-							);
-                		}
+						// Titel gefunden, Daten übernehmen
+						$titel[] = array
+						(
+							'id'          => $objMembers->id,
+							'nachname'    => $objMembers->nachname,
+							'vorname'     => $objMembers->vorname,
+							'titel'       => $titelkey,
+							'datum'       => $objMembers->{$titelkey.'_date'},
+							'turnier'     => '',
+							'link'        => '',
+						);
 					}
 				}
-				
-				// Restliche (nicht markierte) Normen eintragen
-				if(is_array($normen))
+					
+				// Normen suchen
+				$normenMember = unserialize($objMembers->normen); // Normen extrahieren
+				if(is_array($normenMember))
 				{
-					for($x = 0; $x < count($normen); $x++)
+					for($x = 0; $x < count($normenMember); $x++)
 					{
-						if(!$normen[$x]['markiert'])
+						if($normenMember[$x]['title'])
 						{
-							$daten[$normen[$x]['title']]['norm'][] = array
+							$normen[] = array
 							(
+								'id'          => $objMembers->id,
 								'nachname'    => $objMembers->nachname,
 								'vorname'     => $objMembers->vorname,
-								'name'        => $objMembers->vorname.' '.$objMembers->nachname,
-								'norm'        => $normen[$x]['title'],
-								'datum'       => \Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($normen[$x]['date']),
-								'turnier'     => $normen[$x]['tournament'],
-								'link'        => $normen[$x]['url'],
-								'turnierlink' => $normen[$x]['url'] ? '<a href="'.$normen[$x]['url'].'" target="_blank">'.$normen[$x]['tournament'].'</a>' : $normen[$x]['tournament'],
+								'titel'       => $normenMember[$x]['title'],
+								'datum'       => $normenMember[$x]['date'],
+								'turnier'     => trim($normenMember[$x]['tournament']),
+								'link'        => trim($normenMember[$x]['url']),
+								'markiert'    => false
 							);
 						}
+					}	
+				}
+			}
+		}
+
+		// Alle Titel darauf prüfen, welche Norm dazugehören könnte
+		for($x = 0; $x < count($titel); $x++)
+		{
+			$differenz = 999999; // Differenz zwischen Titel- und Normdatum auf hohen Wert setzen
+			// Normen des Spielers und seines Titels suchen
+			for($y = 0; $y < count($normen); $y++)
+			{
+				if($titel[$x]['id'] == $normen[$y]['id'] && $titel[$x]['titel'] == $normen[$y]['titel'])
+				{
+					// Norm markieren
+					$normen[$y]['markiert'] = true;
+					// Mögliche letzte Norm dieses Titels eintragen
+					if($differenz > $titel[$x]['datum'] - $normen[$y]['datum'])
+					{
+						$differenz = $titel[$x]['datum'] - $normen[$y]['datum'];
+						$titel[$x]['turnier'] = $normen[$y]['turnier'];
+						$titel[$x]['link'] = $normen[$y]['link'];
 					}
 				}
 			}
 		}
+
+		// Titel sortieren nach Datum absteigend
+		$titel = \Schachbulle\ContaoHelperBundle\Classes\Helper::sortArrayByFields($titel, array('datum' => SORT_DESC));
+
+		// Markierte Normen entfernen, da bereits in Titeln verwendet
+		$neu = array();
+		for($y = 0; $y < count($normen); $y++)
+		{
+			if(!$normen[$y]['markiert'])
+			{
+				$neu[] = $normen[$y];
+			}
+		}
+		
+		// Normen sortieren nach Datum absteigend
+		$normen = \Schachbulle\ContaoHelperBundle\Classes\Helper::sortArrayByFields($neu, array('datum' => SORT_DESC));
+
+		// Titel übertragen in Ausgabe
+		for($x = 0; $x < count($titel); $x++)
+		{
+			// Nur Titel innerhalb des Zeitfensters
+			if($titel[$x]['datum'] >= $mindate && $titel[$x]['datum'] <= $maxdate)
+			{
+				$daten[$titel[$x]['titel']]['titel'][] = array
+				(
+					'nachname'    => $titel[$x]['nachname'],
+					'vorname'     => $titel[$x]['vorname'],
+					'name'        => $titel[$x]['vorname'].' '.$titel[$x]['nachname'],
+					'norm'        => $titel[$x]['titel'],
+					'datum'       => \Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($titel[$x]['datum']),
+					'turnier'     => $titel[$x]['turnier'],
+					'link'        => $titel[$x]['link'],
+					'turnierlink' => $titel[$x]['link'] ? '<a href="'.$titel[$x]['link'].'" target="_blank">'.$titel[$x]['turnier'].'</a>' : $titel[$x]['turnier'],
+				);
+ 			}
+		}
+		
+		// Normen übertragen in Ausgabe
+		for($x = 0; $x < count($normen); $x++)
+		{
+			// Nur Normen innerhalb des Zeitfensters
+			if($normen[$x]['datum'] >= $mindate && $normen[$x]['datum'] <= $maxdate)
+			{
+				$daten[$normen[$x]['titel']]['norm'][] = array
+				(
+					'nachname'    => $normen[$x]['nachname'],
+					'vorname'     => $normen[$x]['vorname'],
+					'name'        => $normen[$x]['vorname'].' '.$normen[$x]['nachname'],
+					'norm'        => $normen[$x]['titel'],
+					'datum'       => \Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($normen[$x]['datum']),
+					'turnier'     => $normen[$x]['turnier'],
+					'link'        => $normen[$x]['link'],
+					'turnierlink' => $normen[$x]['link'] ? '<a href="'.$normen[$x]['link'].'" target="_blank">'.$normen[$x]['turnier'].'</a>' : $normen[$x]['turnier'],
+				);
+ 			}
+		}
+
+		// Ausgabe
 		$this->Template->daten = $daten;
 
 	}